Refactor GenomicRDD.writeTextRdd to util.TextRddWriter.
