+++ adam-cli/src/main/scala/org/bdgenomics/adam/cli/MergeShards.scala
@@ -19,7 +19,7 @@ package org.bdgenomics.adam.cli
 
 import org.apache.hadoop.fs.Path
 import org.apache.spark.SparkContext
-import org.bdgenomics.adam.rdd.FileMerger
+import org.bdgenomics.adam.util.FileMerger
 import org.bdgenomics.utils.cli._
 import org.kohsuke.args4j.{ Argument, Option => Args4jOption }
